Updated Memcached adapter to check if it can connect to servers before performing any operations

+++ src/Adapters/Memcached.php
@@ -6,7 +6,7 @@ namespace CacheExchange\Adapters;
  * Uses PHP Memcached extension
  * http://www.php.net/manual/en/class.memcached.php
  */
-class Memcached implements \CacheExchange\Interfaces\DatastoreInterface
+class Memcached implements \CacheExchange\Interfaces\Datastore
 {
   /**
    * Memcached object
@@ -14,43 +14,88 @@ class Memcached implements \CacheExchange\Interfaces\DatastoreInterface
    */
   protected $cache;
 
+  protected $ready = false;
+
   /**
-   * @param $servers
+   * __construct
+   * @param array $settings Array of settings
+   *                        "connections" => array of array of settings (multiple servers)
    */
-  public function __construct($servers = [])
+  public function __construct($settings)
   {
     $this->cache = new \Memcached();
-    $this->cache->addServers($servers);
+    $this->addServers($settings["connections"]);
+
+    if ($this->cache->getStats() === false) {
+      // cannot connect to servers
+      return false;
+    } else {
+      // can connect to servers
+      $this->ready = true;
+    }
   }
 
-  public function set(string $key, mixed $value, int $ttl = 0): bool
+  protected function addServers($connections)
   {
-    return $this->cache->set($key, $value, $ttl);
+    // make sure servers aren't being duplcated
+    $existingServers = $this->cache->getServerList();
+    if (!empty($existingServers)) return;
+
+    $this->cache->addServers($connections);
   }
 
-  public function get(string $key): mixed
+  public function store($key, $value, $seconds)
   {
+    if (!$this->ready) {
+      return false;
+    }
+
+    return $this->cache->set($key, $value, $seconds);
+  }
+
+  public function fetch($key)
+  {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->get($key);
   }
 
-  public function exists(string $key): bool
+  public function exists($key)
   {
-    $data = $this->get($key);
+    if (!$this->ready) {
+      return false;
+    }
+
+    $data = $this->fetch($key);
     return !empty($data);
   }
 
-  public function delete(string $key): bool
+  public function delete($key)
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->delete($key);
   }
 
-  public function clear(): bool
+  public function clear()
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->flush();
   }
 
-  public function getKeys(): array|false
+  public function getKeys()
   {
+    if (!$this->ready) {
+      return false;
+    }
+
     return $this->cache->getAllKeys();
   }
 }
